[修改] 解决 token -> UserContext 时，List<Role> 转换异常的问题。

+++ src/main/java/com/microtomato/hirun/modules/system/service/impl/AuthServiceImpl.java
@@ -247,9 +247,9 @@ public class AuthServiceImpl implements IAuthService {
             List<Role> roles = new ArrayList<>();
 
             for (Map map : list) {
-                Long id = (Long) map.get("id");
+                Integer id = (Integer) map.get("id");
                 String name = (String) map.get("name");
-                Role role = new Role(id, name);
+                Role role = new Role(Long.valueOf(id), name);
                 roles.add(role);
             }
